User-configurable exported image size (#14)
User-configurable exported image size

Allow users to specify width & height of exported images (in inches, cm
or pixels). If this option is not enabled, then the window/figure
geometry will be used.

The figure is rendered to an image by creating a new slice view
instance that reuses the individual views from the on-screen rendered
figure. It therefore inherits current inline, crossline, zoom and layout
as what's rendered on screen - the only proper modifications are w.r.t.
geometry.